add all steps of the fourth homework

+++ controllers/contacts/removeById.js
@@ -2,7 +2,8 @@ const { Contact } = require('../../model')
 
 const removeById = async (req, res) => {
   const { id } = req.params
-  const result = await Contact.findByIdAndRemove(id)
+  const { _id } = req.user
+  const result = await Contact.findByIdAndRemove({ owner: _id, _id: id })
   res.json({
     status: 'success',
     code: 200,
